+++ src/main/java/com/fuglkrig/server/classes/EntityCanonBall.java
@@ -10,7 +10,7 @@ public interface EntityCanonBall {
     public void tick();
     public Rectangle getBounds();
 
-    public double getX();
-    public double getY();
+    public int getX();
+    public int getY();
 
 }
